backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.133.0",
    "bcrypt>=3.2.2,<4.0.0",
    "google-genai>=0.6.0",
    "passlib[bcrypt]>=1.7.4",
    "pdfplumber>=0.11.9",
    "pymongo>=4.16.0",
    "python-jose[cryptography]>=3.3.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn>=0.41.0",
]
